--- Runtime/RosMsgs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f6736a9ca092744db117702bb054b10
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
