+++ Runtime/RosMsgs/IRosMsg.cs
@@ -1,0 +1,7 @@
+﻿namespace std_msgs
+{
+    public interface IRosMsg
+    {
+        public string GetRosType();
+    }
+}
