+++ Runtime/RosMsgs/Std/std_msgs.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c8ec94515b014085ab4dda319cff42c3
+timeCreated: 1725971230
